chore: use tsdx and jest for tests
