--- .istanbul.yml
@@ -1,2 +1,0 @@
-instrumentation:
-    excludes: ['**/vendor/**']
