+++ package.json
@@ -21,13 +21,11 @@
     "build:js": "rollup -c",
     "build:types": "tsc --emitDeclarationOnly",
     "build": "npm run clean && npm run build:types && npm run build:js",
-    "lint": "eslint lib test --ext '.ts,.js'",
+    "lint": "eslint src test --ext '.ts,.js'",
     "pretest": "npm run lint",
-    "test": "npm run test:cover",
-    "test:ci-emulate": "trevor",
-    "test:cover": "babel-node --extensions .ts --extensions .js ./node_modules/istanbul/lib/cli.js cover ./test/runner | tap-spec",
-    "test:unit": "babel-node --extensions .ts --extensions .js ./test/runner | tap-spec",
-    "test:debug": "babel-node --extensions .ts --extensions .js debug ./test/runner",
+    "test": "tsdx test",
+    "test:watch": "tsdx test --watch",
+    "test:cover": "tsdx test --coverage",
     "browser-coverage": "npm run test:cover && opener coverage/lcov-report/index.html",
     "prepublish": "in-publish && npm run build || not-in-publish",
     "semantic-release": "semantic-release",
@@ -48,7 +46,6 @@
     "qs": "^6.9.4"
   },
   "devDependencies": {
-    "@types/qs": "^6.9.5",
     "@babel/cli": "^7.12.8",
     "@babel/core": "^7.12.9",
     "@babel/node": "^7.12.6",
@@ -56,13 +53,13 @@
     "@babel/preset-typescript": "^7.12.7",
     "@babel/register": "^7.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
+    "@types/qs": "^6.9.5",
     "@typescript-eslint/eslint-plugin": "2.34.0",
     "@typescript-eslint/parser": "2.34.0",
     "axios": "^0.21.0",
     "axios-mock-adapter": "^1.15.0",
     "babel-eslint": "^10.1.0",
     "babel-plugin-rewire-ts": "1.4.0",
-    "blue-tape": "^1.0.0",
     "codecov": "^3.0.4",
     "cz-conventional-changelog": "^3.1.0",
     "eslint": "^7.2.0",
@@ -75,7 +72,6 @@
     "eslint-plugin-standard": "^5.0.0",
     "husky": "^4.3.0",
     "in-publish": "^2.0.0",
-    "istanbul": "^1.0.0-alpha.2",
     "lint-staged": "^10.5.2",
     "mkdirp": "^1.0.4",
     "moxios": "^0.4.0",
@@ -86,11 +82,9 @@
     "rollup": "^2.34.0",
     "rollup-plugin-babel": "^4.4.0",
     "semantic-release": "^17.0.7",
-    "sinon": "^9.0.2",
-    "tap-spec": "^5.0.0",
-    "trevor": "^2.3.0",
+    "tsdx": "^0.14.1",
     "tslib": "^2.0.3",
-    "typescript": "4.1.2"
+    "typescript": "3.9.3"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
