+++ src/enforce-obj-path.ts
