+++ src/to-plain-object.ts
