+++ test/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: '../.eslintrc.js',
+  globals: {
+    describe: true,
+    it: true,
+    expect: true,
+  },
+}
