+++ test/unit/create-http-client-test.js
@@ -3,7 +3,7 @@ import sinon from 'sinon'
 
 import createHttpClient, {
   __RewireAPI__ as createHttpClientRewireApi,
-} from '../../lib/create-http-client'
+} from '../../src/create-http-client'
 
 import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
