+++ tsconfig.json
@@ -10,6 +10,6 @@
     "esModuleInterop": true,
     "noImplicitThis": false
   },
-  "include": ["lib/**/*", "global.d.ts"],
+  "include": ["src/**/*", "global.d.ts"],
   "exclude": ["node_modules", "dist"]
 }
